// packages/entities/entities-plugins/src/components/free-form/plugins/datakit/types.ts
export interface DatakitUIData {
  /**
   * UI nodes that store data like input/output fields, positions, and other metadata.
   */
  nodes?: UINode[]
  groups?: UIGroup[]
}
/**
 * The phase of the node in the request/response cycle.
 *
 * Note: Phases of implicit nodes are hardcoded and cannot be changed.
 */
export type NodePhase = 'request' | 'response'

/************************************************
 *             Editor global store              *
 ************************************************/

export type NodeId = `node:${number}`
export type EdgeId = `edge:${number}`
export type FieldId = `field:${number}`
export type GroupId = `${NodeId}:${BranchName}`
export type GroupName = `${NodeName}:${BranchName}`

export type IdConnection = NodeId | `${NodeId}.${FieldId}`

export interface NodeField {
  id: FieldId
  name: FieldName
}

export type UINode = {
  name: NodeName
  phase: NodePhase
  position: XYPosition
  fields: { input?: FieldName[], output?: FieldName[] }
  expanded: { input?: boolean, output?: boolean }
  /** hidden in UI */
  hidden?: boolean
}

export interface NodeInstance {
  id: NodeId
  type: NodeType
  name: NodeName
  phase: NodePhase
  position: XYPosition
  expanded: { input?: boolean, output?: boolean }
  fields: { input: NodeField[], output: NodeField[] }
  /** hidden in UI */
  hidden?: boolean
  config?: Record<string, unknown>
}

export interface ConfigEdge {
  source: NodeName
  sourceField?: FieldName
  target: NodeName
  targetField?: FieldName
}

export interface EdgeData {
  source: NodeId
  sourceField?: FieldId
  target: NodeId
  targetField?: FieldId
}

export interface EdgeInstance extends EdgeData {
  id: EdgeId
}

export interface GroupInstance {
  id: GroupId
  ownerId: NodeId
  branch: BranchName
  phase: NodePhase
  memberIds: NodeId[]
  position?: XYPosition
  dimensions?: Dimensions
}

export type FlowGroupNodeData = GroupInstance & {
  memberIds: NodeId[]
}

export type BranchEdgeData = {
  type: 'branch'
  ownerId: NodeId
  branch: BranchName
  groupId: GroupId
}

export type FlowEdge = Edge<EdgeData | BranchEdgeData>

export interface GroupLayout {
  position: XYPosition
  dimensions: Dimensions
  commit: boolean
}

export interface UIGroup {
  name: GroupName
  position: XYPosition
}

export interface EditorState {
  nodes: NodeInstance[]
  edges: EdgeInstance[]
  groups: GroupInstance[]

  /**
   * Whether to schedule an `autoLayout` after the current state is rendered.
   */
  pendingLayout?: false | 'clearHistory' | 'keepHistory'

  /**
   * Whether to schedule a `fitView` after the current state is rendered.
   */
  pendingFitView?: boolean

  cacheConfig?: CacheConfigFormData | null
}

export interface MakeNodeInstancePayload {
  type: NodeType
  name?: NodeName
  phase?: NodePhase
  position?: XYPosition
  fields?: { input?: FieldName[], output?: FieldName[] }
  /** hidden in UI */
  hidden?: boolean
  config?: Record<string, unknown>
}

export interface CreateNodePayload extends MakeNodeInstancePayload {
  type: ConfigNodeType
}

export type DragAction = 'create-node'

export interface DragPayload {
  action: DragAction
  data: {
    type: ConfigNodeType
    anchor: {
      ratioX: number
      ratioY: number
      offsetX: number
      offsetY: number
    }
  }
}

export type DatakitPluginData = FreeFormPluginData<DatakitConfig, DatakitUIData>

export type CacheConfigFormData = CacheConfig & {
  partial_id?: string
}